Exclude Git options from clone step
Should be low-hanging fruit to add this back in the correct way, but for
now, this is the easy fix. If I'm parsing the previous syntax correctly,
the problem was that the repo dir and maybe some segment(s) of the repo
URL are duplicated when cloning... just need to package up the options
excluding those duped args.