Convention: use initial underscores for variable defaults
Variable names that don't start with an underscore should be construed
as potentially exposed to the user (providing them with a simple,
readable override mechanism).
If it starts with an underscore, it's for internal consumption, and
stands in for a value that might otherwise be overridden by user input.

This comes with the added benefit of making the user repo configs a
little less messy.